requirements.txt
playwright>=1.40.0
playwright-stealth>=1.0.6
beautifulsoup4>=4.12.0
duckduckgo-search>=5.0.0
openai>=1.0.0
python-dotenv>=1.0.0
